Use Renjin instead of RServe
Renjin is a complete R implementatin in Java. They provide a JSR223
script engine, which is preferrable to connecting to a running RServe
instance.

This represents a new direction for R scripting in Java. Given that this
is a Java re-implementation of R, this project will from now on be
"scripting-renjin" - allowing for the possibility of a true scripting-r
implementation in the future.