Rename Renjin classes to have "renjin" in the name
The package is now org.scijava.plugins.scripting.renjin.
The classes are now Renjin<Foo> instead of R<Foo>.

This is consistent with other scripting languages such as Jython and
JRuby. It is important in case we ever introduce a second R-based
script language, such as one that uses RServe instead of Renjin.